Various bug fixes relating to new project loading system:  - Delayed project load so it triggers on the sim thread instead of the Mono thread  - Added background area for the recent projects list in Project Window  - On project unload close all editor widgets  - Project Window size is now initialized before attempting to retrieve GUI element bounds  - ScriptLibrary::reload now only performs assembly refresh if script assemblies were previously loaded, otherwise it just additively loads them  - Fixed a race condition due to compiler variable reordering when initializing core versions of certain systems  - Script interop GUI element objects are now properly deleted when their managed instance goes out of scope  - When inputting characters ignore special char symbols generated by ctrl+key presses  - Fixed an issue where during assembly refresh new script objects were being generated, breaking iterators  - Fixed invalid C++/CLR interop methods for HString  - Fixed static order initialization issue - project library internal path now holds the proper value

+++ SBansheeEngine/Source/BsScriptGUILayout.cpp
@@ -12,8 +12,8 @@
 
 namespace BansheeEngine
 {
-	ScriptGUILayout::ScriptGUILayout(MonoObject* instance, GUILayout* layout)
-		:TScriptGUIElementBase(instance, layout), mLayout(layout), mIsDestroyed(false)
+	ScriptGUILayout::ScriptGUILayout(MonoObject* instance, GUILayout* layout, bool ownsNative)
+		:TScriptGUIElementBase(instance, layout), mLayout(layout), mIsDestroyed(false), mOwnsNative(ownsNative)
 	{
 
 	}
@@ -39,7 +39,8 @@ namespace BansheeEngine
 
 			destroyChildren();
 
-			GUILayout::destroy(mLayout);
+			if (mOwnsNative)
+				GUILayout::destroy(mLayout);
 
 			mLayout = nullptr;
 			mIsDestroyed = true;
@@ -200,7 +201,7 @@ namespace BansheeEngine
 	MonoObject* ScriptGUIPanel::createFromExisting(GUIPanel* panel)
 	{
 		MonoObject* managedInstance = metaData.scriptClass->createInstance();
-		ScriptGUILayout* nativeInstance = new (bs_alloc<ScriptGUILayout>()) ScriptGUILayout(managedInstance, panel);
+		ScriptGUILayout* nativeInstance = new (bs_alloc<ScriptGUILayout>()) ScriptGUILayout(managedInstance, panel, false);
 
 		return managedInstance;
 	}
